Move middleware and tests from mozorg to base app
