+++ bedrock/base/middleware.py
@@ -13,6 +13,7 @@ import base64
 import contextlib
 import inspect
 import time
+from email.utils import formatdate
 from functools import wraps
 
 from django.conf import settings
@@ -20,6 +21,7 @@ from django.core.exceptions import MiddlewareNotUsed
 from django.http import Http404, HttpResponse, HttpResponseRedirect
 from django.middleware.locale import LocaleMiddleware as DjangoLocaleMiddleware
 from django.utils import translation
+from django.utils.cache import add_never_cache_headers
 from django.utils.deprecation import MiddlewareMixin
 from django.utils.translation import trans_real
 
@@ -324,3 +326,83 @@ class CSPMiddlewareByPathPrefix(RateLimitedCSPMiddleware):
                     break
 
         return super().process_response(request, response)
+
+
+class CacheMiddleware:
+    def __init__(self, get_response):
+        self.get_response = get_response
+
+    def __call__(self, request):
+        response = self.get_response(request)
+        return self.process_response(request, response)
+
+    def process_response(self, request, response):
+        if request.method not in ("GET", "HEAD"):
+            return response
+        if response.streaming or "cache-control" in response:
+            return response
+
+        # Expiration time in seconds, now + 10 minutes.
+        stamp = int(time.time() + 600)
+
+        response["Cache-Control"] = "max-age=600"
+        response["Expires"] = formatdate(timeval=stamp, localtime=False, usegmt=True)
+        return response
+
+
+class ClacksOverheadMiddleware:
+    # bug 1144901
+
+    def __init__(self, get_response):
+        self.get_response = get_response
+
+    def __call__(self, request):
+        response = self.get_response(request)
+        return self.process_response(request, response)
+
+    @staticmethod
+    def process_response(request, response):
+        if response.status_code == 200:
+            response["X-Clacks-Overhead"] = "GNU Terry Pratchett"
+        return response
+
+
+class HostnameMiddleware:
+    def __init__(self, get_response):
+        if not settings.ENABLE_HOSTNAME_MIDDLEWARE:
+            raise MiddlewareNotUsed
+
+        values = [getattr(settings, x) for x in ["HOSTNAME", "CLUSTER_NAME"]]
+        self.backend_server = ".".join(x for x in values if x)
+
+        self.get_response = get_response
+
+    def __call__(self, request):
+        response = self.get_response(request)
+        return self.process_response(request, response)
+
+    def process_response(self, request, response):
+        response["X-Backend-Server"] = self.backend_server
+        return response
+
+
+class VaryNoCacheMiddleware:
+    def __init__(self, get_response):
+        if not settings.ENABLE_VARY_NOCACHE_MIDDLEWARE:
+            raise MiddlewareNotUsed
+        self.get_response = get_response
+
+    def __call__(self, request):
+        response = self.get_response(request)
+        return self.process_response(request, response)
+
+    @staticmethod
+    def process_response(request, response):
+        if "vary" in response:
+            path = request.path
+            if path != "/" and not any(path.startswith(x) for x in settings.VARY_NOCACHE_EXEMPT_URL_PREFIXES):
+                del response["vary"]
+                del response["expires"]
+                add_never_cache_headers(response)
+
+        return response
